requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
pyjwt==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
